detail한 수정작업

+++ src/views/Event.vue
@@ -72,14 +72,12 @@
             </template>
             <div >
               <div class="rounded shadow bg-white">
-                <b-dropdown-item v-on:click="onClickEditEvent">edit</b-dropdown-item>
+                <b-dropdown-item v-if="isActivated" v-on:click="onClickEditEvent">edit</b-dropdown-item>
                 <b-dropdown-item v-on:click="onClickDestroyEvent">delete</b-dropdown-item>
-                <b-dropdown-item v-on:click="onClickCloseEvent">close</b-dropdown-item>
-                <b-dropdown-item v-b-modal.modal-prevent-closing>receive</b-dropdown-item>
+                <b-dropdown-item v-if="isActivated" v-on:click="onClickCloseEvent">close</b-dropdown-item>
+                <b-dropdown-item v-if="isReceived === false" @click="clickModal">receive</b-dropdown-item>
                 <b-modal
-                  id="modal-prevent-closing"
-                  ref="modal"
-                  @show="resetModal"
+                  ref="receive-deposit-modal"
                   @hidden="resetModal"
                   @ok="handleOk"
                 >
@@ -165,7 +163,8 @@ export default {
       userType: '',
       bankCode: '',
       accountNumber: '',
-      nameState: ''
+      nameState: '',
+      isReceived: ''
     }
   },
   methods: {
@@ -186,17 +185,34 @@ export default {
       this.eventHash = data.event_hash
       this.userId = data.user_id
       this.userType = data.userType
+      this.isReceived = data.is_received
     },
     onClickEditEvent () {
       this.$router.replace({ path: '/event/edit/' + this.$route.params.id }).catch(() => {})
     },
     async onClickDestroyEvent () {
-      await API.destroyEventAPI(this.$http, this.$env.apiUrl, this.$route.params.id)
+      if (this.isActivated) {
+        alert('종료된 이벤트만 삭제할 수 있습니다.')
+        return
+      }
+      if (confirm('이벤트 삭제 시 모든 이체 내역을 조회할 수 없게 됩니다. 정말 삭제하시겠습니까?') === false) {
+        return
+      }
+      const res = await API.destroyEventAPI(this.$http, this.$env.apiUrl, this.$route.params.id)
+      if (res.data.result === errorcode.UNKNOWN_ERROR) {
+        alert(res.data.detail)
+        return
+      }
       this.$router.replace({ path: '/' }).catch(() => {})
     },
     async onClickCloseEvent () {
-      await API.closeEventAPI(this.$http, this.$env.apiUrl, this.$route.params.id)
-      this.$router.replace({ path: '/' }).catch(() => {})
+      if (confirm('이벤트 종료 시 송금 링크가 비활성화 됩니다. 이벤트를 종료하시겠습니까?') === false) {
+        return
+      }
+      const res = await API.closeEventAPI(this.$http, this.$env.apiUrl, this.$route.params.id)
+      if (res.data.result === errorcode.SUCCESS) {
+        this.isActivated = false
+      }
     },
     onClickBreakdownEvent () {
       this.$router.replace({ path: `/event/${this.$route.params.id}/breakdown` }).catch(() => {})
@@ -243,6 +259,7 @@ export default {
       if (res.data.result !== errorcode.SUCCESS) {
         alert(res.data.detail)
       }
+      this.$refs['receive-deposit-modal'].hide()
     },
     getTransferUrl () {
       return `${this.$env.hostUrl}/fin/transfer/${this.eventHash}`
@@ -255,7 +272,17 @@ export default {
       this.nameState = valid
       return valid
     },
+    clickModal () {
+      if (this.isActivated) {
+        console.log('hi')
+        alert('종료 이후에만 출금할 수 있습니다.')
+        return
+      }
+      this.$refs['receive-deposit-modal'].show()
+    },
     resetModal () {
+      this.bankCode = ''
+      this.accountNumber = ''
     },
     handleOk (bvModalEvt) {
       // Prevent modal from closing
